Updates readResult() to 16 bit integer
Used for compatibility across boards and to match .cpp changes

+++ libraries/OPT3001/OPT3001.h
@@ -17,7 +17,7 @@ class opt3001
 {
   public:
 	void begin();
-	uint32_t readResult();
+	uint16_t readResult();
 	uint16_t readManufacturerId();
 	uint16_t readDeviceId();
 	uint16_t readConfigReg();
